api/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
pymongo==4.3.3
Werkzeug==2.2.3
python-dotenv==0.21.1
dnspython==2.3.0
